Ignore insignificant request headers for stub_request
On Travis CI the User-Agent header was slightly different, causing a
failed test. Since those headers don't impact the request, they can be
removed. I'm leaving the tracker token for now as it is tied to the ENV
var and would have an impact on the request. There is probably a better
way to test that, but this will do for now.